Add copyright and licensing boilerplate text
The library is licensed according to the revised (3-clause) BSD license,
without an "advertising clause".

There was some consideration given to licensing utilities/* according to
the GNU GPL, but in the end they're trivial enough that the gain (distributed
improvements are *required* to be contributed, rather than just encouraged)
doesn't outweigh the added complexity.