Move headers from sam/* to cansam/*
We can't really assume that we'll be the only SAM-related C++ library
someone might have installed, so install headers in a more specific
place.  And change the source tree to reflect that, so uninstalled
include paths continue to reflect corresponding installed ones.

Also introduces cansam/sam/* for truly SAM-related headers, so that
cansam/* is available for ancillary useful things.