Fix buffering bugs exposed when reading from pipes
Various code (stupidly) assumed that read() would more or less fill
the buffer supplied (except at EOF).  This is of course incorrect for
pipes and sockets.

Most entertainingly, flush_make_space() would double its buffer on each
call, assuming it was being called because little space was available.
When it is being called because the pipe is providing only a few bytes at
a time (thus leaving much space available), this soon exhausted memory.
(The magic numbers introduced correspond to not reallocating when there
is a "reasonable" amount available, increasing somewhat when the buffer
has increased in capacity.)