Take advantage of recent samstream changes
In particular, failbit|badbit exceptions are already set by default,
and constructors throw an exception themselves on open failures.

This changes the behaviour of samcat when reading fails: previously it
would fail immediately with an error message, now it prints a message,
continues with any subsequent input files, and then fails.

Introduces a Boost dependency (scoped_array) for samsplit.