Tidy-ups for compatibility with recent compilers
In internal code, I would just include a C-style header or add a using
directive to continue writing plain ptrdiff_t, but a public header needs
to be more conservative.  There is likely a similar problem w.r.t. size_t
still in the code.

Having to forward declare the decimal() templates is not great, and both
this code and the alignment & header iterators should be revisited.