Avoid catching std::ios_base::failure to work around GCC bug
Work around for <https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66145>.
In GCC 5 and 6 there are C++98 and C++11 flavours of ios_base::failure
and it is unclear how to arrange to catch the one libstdc++ is actually
throwing without excess compiler-specific verbiage.

This try-catch is really just there to catch anything that setstate()
throws (which is always std::ios_base::failure), so just write it as
`catch (...)` as the only thing in the try block is setstate() itself.